+++ src/Datasets/Generators/Blob.php
@@ -42,11 +42,11 @@ class Blob implements Generator
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(array $center = [0.0, 0.0], $stddev = 1.)
+    public function __construct(array $center = [0., 0.], $stddev = 1.)
     {
         $d = count($center);
 
-        if ($d === 0) {
+        if ($d < 1) {
             throw new InvalidArgumentException('Cannot generate data of less'
                 . ' than 1 dimension.');
         }
@@ -56,24 +56,24 @@ class Blob implements Generator
         }
 
         if ($d !== count($stddev)) {
-            throw new InvalidArgumentException('The number of center'
-                . ' coordinates and standard deviations must be equal.');
+            throw new InvalidArgumentException("The number of center"
+                . " coordinates and standard deviations must the same.");
         }
 
         foreach ($center as $column => $mean) {
             if (!is_int($mean) and !is_float($mean)) {
                 throw new InvalidArgumentException('Center coordinate must be'
-                    . ' an integer or float.');
+                    . ' an integer or float, ' . gettype($mean) . ' found.');
             }
 
             if (!is_int($stddev[$column]) and !is_float($stddev[$column])) {
                 throw new InvalidArgumentException('Standard deviation must be'
-                    . ' an integer or float.');
+                    . ' an integer or float, '. gettype($mean) . ' found.');
             }
 
             if ($stddev[$column] <= 0) {
-                throw new InvalidArgumentException('Standard deviation must be'
-                 . ' greater than 0.');
+                throw new InvalidArgumentException("Standard deviation must be"
+                 . " greater than 0, $stddev[$column] given.");
             }
         }
 
@@ -100,8 +100,8 @@ class Blob implements Generator
     public function generate(int $n = 100) : Dataset
     {
         $samples = Matrix::gaussian($n, $this->dimensions())
-            ->add($this->center)
             ->multiply($this->stddev)
+            ->add($this->center)
             ->asArray();
 
         return Unlabeled::quick($samples);