+++ src/Datasets/Generators/HalfMoon.php
@@ -60,18 +60,18 @@ class HalfMoon implements Generator
                                 float $noise = 0.1)
     {
         if ($scale < 0.) {
-            throw new InvalidArgumentException('Scaling factor must be greater'
-                . ' than 0.');
+            throw new InvalidArgumentException("Scaling factor must be greater"
+                . " than 0, $scale given.");
         }
 
         if ($rotation < 0. or $rotation > 360.) {
-            throw new InvalidArgumentException('Rotation must be between 0 and'
-                . ' 360 degrees.');
+            throw new InvalidArgumentException("Rotation must be between 0 and"
+                . " 360 degrees, $rotation given.");
         }
 
-        if ($noise <= 0. or $noise > 1.) {
-            throw new InvalidArgumentException('Noise factor must be great than'
-                . ' 0 and less than or equal to 1.');
+        if ($noise < 0. or $noise > 1.) {
+            throw new InvalidArgumentException("Noise factor must be between 0"
+                . " and less 1, $noise given.");
         }
 
         $this->center = Vector::quick([$x, $y]);
@@ -108,8 +108,8 @@ class HalfMoon implements Generator
         $samples = Matrix::fromVectors([$r->cos(), $r->sin()])
             ->transpose()
             ->add($noise)
-            ->add($this->center)
             ->multiply($this->scale)
+            ->add($this->center)
             ->asArray();
 
         return Unlabeled::quick($samples);