+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - Min Max Normalizer is now Elastic
     - TF-IDF Transformer is now Elastic
     - Added Huber Loss cost function
+    - Added Swiss Roll generator
     - Moved Generators to the Datasets directory
     - Added Persister interface for Persistable objects
     - Added overwrite protection to Persistent Model meta estimator